Update #2
I have changed the serialization format of the .3ds to binary, currently
i need to change the loading code of the format to load all of the
models currently in the scene, for now it opens the first model in the
file. Also chose to add the executables so you guys can try it out
easier :D Enjoy.