Minor Update #4
Just fixed the bug on loading so now you can load a .3ds scene
perfectly, without loading to few or too many objects into the scene
(the bug originally consisted of being able to only load the first
object inthe .3ds file and not all of them! then i found a way to load
all of them, but it turned out to always load just one more object than
in the files!) so it is now considered fixed!