Update #5
Changed the entire folder structure so that i will hopefully not get
into trouble legally.
Also i have optimized the 3D scene window / pane to be able to contain
approx. 5.38 times more polygons!
And last but not least, now you can also use the project explorer so
when you're selecting an object from the explorer, then it selects that
particular object in the 3d scene! And the objects also have a boundary
box to see which is selected!